+++ src/app/app.module.ts
@@ -2,9 +2,9 @@ import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { ListarComponent } from './Persona/toList/listar.component';
-import { AgregarComponent } from './Persona/add/agregar.component';
-import { EditarComponent } from './Persona/edit/editar.component';
+import { ListarComponent } from './components/toList/listar.component';
+import { AgregarComponent } from './components/add/agregar.component';
+import { EditarComponent } from './components/edit/editar.component';
 import { FormsModule } from '@angular/forms';
 import{ ServiceService } from '../app/Service/service.service';
 import {HttpClientModule} from '@angular/common/http';
